refactor: move remaining tui db calls into core/accounts, core/rules, core/trends
- core/accounts.ts: account CRUD, balance history, manual asset creation, CSV import, dupe deletion
- core/rules.ts: category rule/name rule upsert+apply, category CRUD (create/rename/delete/flexibility)
- core/trends.ts: buildTrendViews, generateAllPeriods, getPeriodTotals with all period SQL
- core/dateUtils.ts: export MONTHS so core/trends.ts can use it without a tui dependency
- tui/Accounts.tsx, Rules.tsx, Trends.tsx: now import from core; no more direct db.prepare calls in tui
